restructure: auth 폴더 제거
