+++ pages/login/oauth2/callback/google.tsx
@@ -1,4 +1,4 @@
-export { GoogleSocialLoginRedirect as default } from '@/pages/auth/login/oauth2/callback/google';
+export { GoogleSocialLoginRedirect as default } from '@/pages/login/oauth2/callback/google';
 
 // 기존 export default를 export로 변경했습니다.
